Command handling fix

+++ commands/grant.js
@@ -14,7 +14,7 @@ module.exports = {
 				const member = message.mentions.members.first();
 				if (member != null && user != null && !member.roles.cache.some(role => role.id == mainRoleID)) {
 					member.roles.add(`${mainRoleID}`);
-					return message.delete().then(message.channel.send({ content: `${user} now has the <@&${mainRoleID}>`, allowedMentions: { parse: [] } }));
+					return message.delete().then(message.channel.send({ content: `${user} now has the <@&${mainRoleID}> role`, allowedMentions: { parse: [] } }));
 				} else {
 					return message.delete().then(message.channel.send('Invalid user specified or user already has role c:'));
 				}
